+++ src/modules/rocket_launch/contracts_window.cpp
@@ -59,15 +59,13 @@ bool planButtonDisabled(const Contract &contract) {
 void acceptContract(flecs::world &world, flecs::entity contractE) {
   Contract &contract = contractE.get_mut<Contract>();
   contract.status = ContractStatus::Accepted;
-  world.get_mut<Company>().balance +=
-      static_cast<int64_t>(contract.upfront_payment);
+  world.get_mut<Company>().balance += contract.upfront_payment;
 }
 
 void rejectContract(flecs::world &world, flecs::entity contractE) {
   Contract &contract = contractE.get_mut<Contract>();
   if (contract.status == ContractStatus::Accepted) {
-    world.get_mut<Company>().balance -=
-        static_cast<int64_t>(contract.upfront_payment);
+    world.get_mut<Company>().balance -= contract.upfront_payment;
   }
   contract.status = ContractStatus::Closed;
   contract.failed = true;
@@ -188,8 +186,7 @@ void drawContractsWindow(flecs::entity winE) {
       ImGui::TextUnformatted(statusStr);
 
       ImGui::TableSetColumnIndex(4);
-      ImGui::Text("%.2f",
-                  contract.upfront_payment + contract.completion_payment);
+      ImGui::Text("%u", contract.upfront_payment + contract.completion_payment);
 
       ImGui::TableSetColumnIndex(5);
       auto targetOrbit = contractE.target<ContractTargetOrbit>();
